Converts to SL-compatible sit targets
- Convert to Secondlife Compatible sit targets
- Add configurable to Halcyon-Sample.ini to allow legacy mode for grids wishing not to break content
- Differentiate three states in sit math
Code currently handles three types of sit target math: SecondLife, Legacy, and None
- Correct information sent in AvatarSitResponse
This now matches SL's response to the same inputs as verified by reviewing the Alchemy Message Log response after setting up identical tests in both grids.

NOTES:

An option is left to switch into the legacy mode for grids who wish to not break content.  This means that this setting can be set per-region, which allows for a range of possibilities.

The choice to default to SL-compat mode was made to facilitate new grids using this codebase: most scripts will be coming from SecondLife and it's really very annoying to have to fix them up - assuming the importer even knows how.

BREAKING CHANGE: This changes an established behavior in the world, causing all Halcyon-specific sit script helpers to break, but allowing the ones coded for SL to work.